requirements.txt
tweepy~=4.14.0
fastapi~=0.110.0
sqlalchemy~=2.0.30
alembic~=1.13.1
openai~=1.30.1
python-dotenv~=1.0.1
asyncpg~=0.29.0
psycopg2-binary~=2.9.9
uvicorn~=0.29.0
requests>=2.26.0
